Nearby strategy improved to accept a range of field radii, making it a lot more useful when we have to guess the field radius. Added versioning to API, split the domain and REST models and established explicit property mapping. Also switched some floats to doubles for more consistency and added a few more nice things here and there.

+++ src/WatneyAstrometry.Core/BlindSearchStrategy.cs
@@ -104,7 +104,7 @@ namespace WatneyAstrometry.Core
                 {
                     int n = 0;
                     //for (float dec = 0; dec <= 90; dec += radius, n++)
-                    float dec = 0;
+                    double dec = 0;
                     bool complete = false;
                     while(!complete)
                     {
